- STM32F 코드 정비중 중간 저장

+++ yss/inc/drv/Clock.h
@@ -47,14 +47,16 @@ class Mainpll
 class Mainpll
 {
   public:
-	bool enable(unsigned char src, unsigned long vcoMhz, unsigned char pDiv, unsigned char qDiv, unsigned char rDiv);
+	bool enable(unsigned char src, unsigned int vcoMhz, unsigned char pDiv, unsigned char qDiv, unsigned char rDiv);
+
 	unsigned int getFreq(void);
+
 };
 
 class Saipll
 {
   public:
-	bool enable(unsigned long vcoMhz, unsigned char pDiv, unsigned char qDiv, unsigned char rDiv);
+	bool enable(unsigned int vcoMhz, unsigned char pDiv, unsigned char qDiv, unsigned char rDiv);
 	unsigned int getLcdFreq(void);
 	unsigned int getSaiFreq(void);
 };
